refactor(favorite): 응원 팀 등록 도메인 분리 (user -> favoriteteam)
