+++ src/main/java/com/sparta/spartatigers/domain/favoriteteam/repository/FavoriteTeamRepository.java
@@ -1,9 +1,9 @@
-package com.sparta.spartatigers.domain.user.repository;
+package com.sparta.spartatigers.domain.favoriteteam.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sparta.spartatigers.domain.team.model.entity.Team;
-import com.sparta.spartatigers.domain.user.model.entity.FavoriteTeam;
+import com.sparta.spartatigers.domain.favoriteteam.model.entity.FavoriteTeam;
 import com.sparta.spartatigers.domain.user.model.entity.User;
 
 public interface FavoriteTeamRepository extends JpaRepository<FavoriteTeam, Long> {
